Add event listener to 'gridButton' to resize grid
The event listener on 'gridButton' calls 'resizeGrid' function which will prompt the user to enter a grid size between 1-100.
If the proper grid size has been entered, the previous grid will be cleared a new grid will be generated using the new grid size.
Otherwise the user will be alerted to enter a number between 1-100.